app/requirements_app.txt
streamlit==1.28.2
joblib== 1.3.2
pandas==2.1.3
Pillow==10.1.0
yfinance==0.2.36
numpy==1.26.2
matplotlib==3.8.2
scikit_learn==1.3.2
keras==2.15.0
tensorflow==2.15.0